Added Trainer Controller and fix some issues in Trainer Repo

+++ src/main/java/fitness_app_be/fitness_app/BusinessLayer/Impl/TrainerServiceImpl.java
@@ -1,13 +1,9 @@
 package fitness_app_be.fitness_app.BusinessLayer.Impl;
 
-
 import fitness_app_be.fitness_app.BusinessLayer.TrainerService;
 import fitness_app_be.fitness_app.Domain.Trainer;
-import fitness_app_be.fitness_app.Domain.User;
 import fitness_app_be.fitness_app.ExceptionHandlingLayer.TrainerNotFoundException;
-import fitness_app_be.fitness_app.ExceptionHandlingLayer.UserNotFoundException;
 import fitness_app_be.fitness_app.MapperLayer.TrainerMapper;
-import fitness_app_be.fitness_app.MapperLayer.UserMapper;
 import fitness_app_be.fitness_app.PersistenceLayer.Entity.TrainerEntity;
 import fitness_app_be.fitness_app.PersistenceLayer.TrainerRepository;
 import lombok.RequiredArgsConstructor;
@@ -22,56 +18,64 @@ public class TrainerServiceImpl implements TrainerService {
 
     private final TrainerRepository trainerRepository;
     private final TrainerMapper trainerMapper;
-    private final UserMapper userMapper;
 
     @Override
-    public List<Trainer> getAllTrainers(){
-        return  trainerRepository.findAll().stream()
+    public List<Trainer> getAllTrainers() {
+        return trainerRepository.findAll().stream()
                 .map(trainerMapper::entityToDomain)
                 .collect(Collectors.toList());
     }
 
     @Override
-    public Trainer getTrainerById(Long id){
+    public Trainer getTrainerById(Long id) {
         return trainerRepository.findById(id)
                 .map(trainerMapper::entityToDomain)
                 .orElseThrow(() -> new TrainerNotFoundException(id));
     }
 
     @Override
-    public Trainer createTrainer(Trainer trainer){
+    public Trainer createTrainer(Trainer trainer) {
         TrainerEntity newEntity = trainerMapper.domainToEntity(trainer);
         TrainerEntity savedTrainerEntity = trainerRepository.save(newEntity);
         return trainerMapper.entityToDomain(savedTrainerEntity);
     }
 
     @Override
-    public void deleteTrainer(Long id){
+    public void deleteTrainer(Long id) {
+        if (!trainerRepository.existsById(id)) {
+            throw new TrainerNotFoundException(id);
+        }
         trainerRepository.deleteById(id);
     }
 
     @Override
-    public Trainer getTrainerByEmail(String email){
+    public Trainer getTrainerByEmail(String email) {
         return trainerRepository.findByEmail(email)
                 .map(trainerMapper::entityToDomain)
                 .orElseThrow(() -> new TrainerNotFoundException("Trainer with email: " + email + " not found"));
     }
 
     @Override
-    public List<Trainer> searchTrainersByPartialUsername(String partialUsername){
+    public List<Trainer> searchTrainersByPartialUsername(String partialUsername) {
         return trainerRepository.findByUsernameContainingIgnoreCase(partialUsername).stream()
                 .map(trainerMapper::entityToDomain)
                 .collect(Collectors.toList());
     }
 
     @Override
-    public Trainer updateTrainer(Trainer trainer){
+    public Trainer updateTrainer(Trainer trainer) {
         TrainerEntity existingTrainerEntity = trainerRepository.findByEmail(trainer.getEmail())
                 .orElseThrow(() -> new TrainerNotFoundException("Trainer with email " + trainer.getEmail() + " not found"));
 
+        // Update all necessary fields
+        existingTrainerEntity.setFirstName(trainer.getFirstName());
+        existingTrainerEntity.setLastName(trainer.getLastName());
+        existingTrainerEntity.setUsername(trainer.getUsername());
+        existingTrainerEntity.setAge(trainer.getAge());
+        existingTrainerEntity.setGender(trainer.getGender());
         existingTrainerEntity.setExpertise(trainer.getExpertise());
-        TrainerEntity updatedTrainerEntity =  trainerRepository.save(existingTrainerEntity);
-        return trainerMapper.entityToDomain(updatedTrainerEntity);
 
+        TrainerEntity updatedTrainerEntity = trainerRepository.save(existingTrainerEntity);
+        return trainerMapper.entityToDomain(updatedTrainerEntity);
     }
 }
